Add --enable-opsmgr flag in omsagent bundle to open port 1270
In order for omsagent to be able to connect to SCOM, port 1270
nees to be opened. A new flag (--enable-opsmgr) has been added
in omsagent bundle for opening port 1270. If someone intends to
use omsagent with SCOM, he will now have an option to pass the
flag during install/upgrade of the agent. Also the port 1270 will
be opened(if not already open) when scom workspace is created.

Following changes are also done:
1. Added scom workpsace in aggregate commands (start all, stop all etc.)
2. Removed code duplicity in timer based plugins.
3. Removed a warning in scom plugin's unittest.